backend/requirements.txt
Django
mysqlclient
djangorestframework
djangorestframework-simplejwt
PyJWT
# python-dateutil
djoser
django-cors-headers
Pillow